# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: countries
    description: Country operations
  - name: recipes
    description: Recipe operations
  - name: search
    description: Search operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /countries:
    get:
      operationId: listCountries
      tags: [countries]
      summary: List all countries
      description: Returns all available countries
      responses:
        "200":
          description: List of countries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Country"

  /countries/{id}:
    get:
      operationId: getCountry
      tags: [countries]
      summary: Get a country with its recipes
      description: Returns a single country and all its associated recipes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Country detail with recipes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CountryDetail"
        "404":
          description: Country not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /recipes/{id}:
    get:
      operationId: getRecipe
      tags: [recipes]
      summary: Get a single recipe
      description: Returns a single recipe by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Recipe detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "404":
          description: Recipe not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /search:
    get:
      operationId: search
      tags: [search]
      summary: Search countries and recipes
      description: Searches countries and recipes by name or description
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Ingredient:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        amount:
          type: string
      required:
        - id
        - name
        - amount

    CookStep:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        instruction:
          type: string
        materials:
          type: array
          items:
            type: string
      required:
        - id
        - title
        - instruction
        - materials

    Recipe:
      type: object
      properties:
        id:
          type: string
        countryId:
          type: string
        region:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
        image:
          type: string
        category:
          type: string
          nullable: true
        prepTime:
          type: string
          nullable: true
        cookTime:
          type: string
          nullable: true
        servings:
          type: integer
          nullable: true
        difficulty:
          type: string
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/Ingredient"
        steps:
          type: array
          items:
            $ref: "#/components/schemas/CookStep"
        culturalNote:
          type: string
          nullable: true
        tips:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - countryId
        - title
        - description
        - image
        - difficulty
        - ingredients
        - steps
        - tips
        - createdAt

    Country:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        flag:
          type: string
        tagline:
          type: string
          nullable: true
        description:
          type: string
        region:
          type: string
          nullable: true
        image:
          type: string
        heroImage:
          type: string
          nullable: true
        cuisineLabel:
          type: string
          nullable: true
        rating:
          type: number
        recipeCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - flag
        - description
        - image
        - rating
        - recipeCount
        - createdAt

    CountryDetail:
      allOf:
        - $ref: "#/components/schemas/Country"
        - type: object
          properties:
            recipes:
              type: array
              items:
                $ref: "#/components/schemas/Recipe"
          required:
            - recipes

    SearchResult:
      type: object
      properties:
        countries:
          type: array
          items:
            $ref: "#/components/schemas/Country"
        recipes:
          type: array
          items:
            $ref: "#/components/schemas/Recipe"
      required:
        - countries
        - recipes

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error